common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
description = "Common utilities for tools"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "certifi>=2024.8.30",
    "click>=8.1.8",
    "feedparser>=6.0.12",
    "gtts>=2.5.4",
    "groq>=0.13.0",
    "langchain>=1.0.5",
    "langchain-google-genai>=3.0.2",
    "langchain-groq>=0.2.1",
    "langchain-openai>=0.2.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=20.0",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
